notebook: Add custom build backend for codegen (#718)
This is built on top of #663, and attempts to address the awkward
interaction between the python build process and the frontend codegen
that we need to do as part of packaging.

The original approach, of putting a guard in `build.rs` has numerous
disadvantages. Chief among them is that it completely breaks building
from sdist, because maturin rearranges our sources files in such a way
that `widget.js` is not where `build.rs` expects it to be. It's also a
great disappointment that `cargo test` and friends now have an
artificial dependency on a frontend codegen step that they don't really
require. And none of this gets us to the point where `uv build` does
what you'd expect - it just yells at the developer to go run `yarn`.

Ideally, we want this all to be baked into the python build system, so
that when you run `uv build`, you get a distribution with the latest
up-to-date frontend assets.

Maturin doesn't offer any hooks for running a codegen step, so this
change wraps maturin in our own custom PEP 517 build backend.
Thankfully, there's a way to do this entirely within tree, so we don't
need a separate package on pypi. The custom build backend only attempts
to do codegen when running in-tree. When building from sdist, it's just
a passthrough to maturin.

I've also updated the CI python workflow. We build the sdist, and then
use that to build wheels. This ensures that sdist builds are working,
and ensures that we're always generating the frontend assets in the same
way. It's a little hacky that we're using PyO3/maturin-action (which
bypasses our build system) for building wheels, but it's safe since our
build system would effectively just call maturin anyway. I tried using
cibuildwheel instead of maturin-action, and it was a bit of a disaster.
And in any case, we need to invoke `maturin build` directly for
cross-compiling the wasm binary, because maturin's PEP 517 backend
doesn't seem support cross compilation.


https://linear.app/foxglove/issue/FG-13367/integrate-front-end-assets-build-into-maturins-workflow