Include ROS distro information in server info message (#716)
In order to support click-to-publish in the 3d panel on a WebSocket
connection, the ServerInfo message needs to include the value of the
ROS_DISTRO environment variable. (Publishing is only supported for ROS
sources right now.) Prior versions of the Bridge included this but the
new releases for ROS2 do not, and the SDK doesn't expose a mechanism for
it.

This PR adds an undocumented `server_info` option to the Rust and C++
server options. This metadata will be included in the ServerInfo message
sent to the app. The only use case today is to convey ROS distro, but we
may extend in the future. I named this "server_info" instead of
"metadata" since it's specifically for information sent to clients, but
am open to better names.

This does not add python support, and we have no workarounds for people
wanting to spoof support for 3d publishing there. ERT-689 covers this
(bringing 3d publish support to non-ROS systems).