Support write metadata in c++ and add it to examples (#715)
### Changelog

Add support to MCAP sink to write MCAP metadata.

### Docs

None

### Description

This builds on Jennifer's work to add support for write metadata to c++.
I also add writing metadata to the mcap examples for all 3 languages,
it's simple enough that I think it's more helpful than harmful, and
doesn't warrant a full standalone example.

I have a chicken and egg problem with the Python example. It needs the
new released version of the library from this PR so it can build. I will
split the change to the example into its own PR so this can be merged
and released first.

---------

Co-authored-by: jenn738 <jenniferjain10@gmail.com>
Co-authored-by: Jennifer Jain <Jenn738@users.noreply.github.com>
Co-authored-by: Eric Lujan <eric@foxglove.dev>