playground: fix condition to allow deploying on main branch (#731)
### Changelog
None

### Docs

None

### Description

Since https://github.com/foxglove/foxglove-sdk/pull/672, the playground
deploy job was always skipped on the main branch because this condition
was false (`github.event.pull_request` is not present on a main branch
run). This PR inverts the condition using a different boolean signal
(`head.repo.fork`) that we already used in the mcap repo.