python: Fix windows wheel builds (#738)
In #690, we worked around a CI issue, but we went about it the wrong
way. By turning off `maturin-action`'s `--find-interpreter`, we ended up
building only a single wheel against python 3.9 for windows x64.

The root of the original problem that #690 was trying to tackle, is that
github's `windows-latest` runner doesn't have _any_ 32-bit pythons
installed. This change uses `setup-python` to install the versions we
care about, and restores the use of `maturin-action`'s
`--find-interpreter` so that we build against them.

```
Archive:  wheels-windows-x86_64.zip
    testing: foxglove_sdk-0.15.2-cp310-cp310-win_amd64.whl   OK
    testing: foxglove_sdk-0.15.2-cp312-cp312-win_amd64.whl   OK
    testing: foxglove_sdk-0.15.2-cp311-cp311-win_amd64.whl   OK
    testing: foxglove_sdk-0.15.2-cp313-cp313-win_amd64.whl   OK
    testing: foxglove_sdk-0.15.2-cp314-cp314-win_amd64.whl   OK
    testing: foxglove_sdk-0.15.2-cp39-cp39-win_amd64.whl   OK

Archive:  wheels-windows-x86.zip
    testing: foxglove_sdk-0.15.2-cp310-cp310-win32.whl   OK
    testing: foxglove_sdk-0.15.2-cp311-cp311-win32.whl   OK
    testing: foxglove_sdk-0.15.2-cp312-cp312-win32.whl   OK
    testing: foxglove_sdk-0.15.2-cp313-cp313-win32.whl   OK
    testing: foxglove_sdk-0.15.2-cp314-cp314-win32.whl   OK
    testing: foxglove_sdk-0.15.2-cp39-cp39-win32.whl   OK
```

Fixes: FG-13419
Fixes: #729