playground: update layout (#748)
### Changelog
None

### Docs
None

### Description
Update the default layout to move raw message panel and plot below the
3d panel. I think it looks nicer to have the 3d panel have more space.

---
Side note: it was annoying to make this change because I had to manually
change `force: true` in the code to actually have the layout override
the local storage variant. Took me a moment to realize that.