rust: Enable rust-2024-compatibility lints (#763)
### Changelog
None

### Docs
None

### Description
Enables lint rules to ease migration to the 2024 edition.

This change disables several lints that don't benefit us until we're
actually ready to cut over to the 2024 edition, at which point they'll
be useful for flagging areas for a one-time manual inspection.

This is the spiritual successor to #248.