Create WebSocket server handler for PlaybackControlRequest (#735)
### Changelog
None

### Docs
None

### Description
Add in Rust SDK functionality for the messages introduced in #728;
namely:
- Introduce the `RangedPlayback` capability, which indicates that a
server connected over WebSocket is providing data from a fixed playback
time range.
- Introduce the `playback_time_range` option in `ServerOptions`, where
the server is expected to declare the time range of the data it is
playing back. This option will implies the `RangedPlayback` capability
and will enable it if it is not already enabled
- Add the `on_playback_control_request` function to the `ServerListener`
trait, which allows a server to receive playback control requests from
the client. It requires the server to send a `PlaybackState` message in
response.
- Add the `broadcast_playback_state` function, which allows a server to
broadcast a change in its playback state in response to anything other
than a `PlaybackControlRequest` message (for example, reaching the end
of playback or an external process pausing playback).

These changes are currently guarded behind the `unstable` FF and
excluded from documentation.