rust: specify options when creating mcap writer from context (#774)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

- rust: add the `Context::mcap_writer_with_options` method

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

This PR adds a method on the `Context` that makes it easier to create an
MCAP writer with options (such as seeking disabled). I'm adding this
method because the `.context(..)` method is hidden from public docs and
it is more ergonomic to create the writer off the context.

Fixes FIRE-223