C/C++: allow custom write implementations for MCAP writer (#764)
### Changelog
- C++: added CustomWriter to allow C++ clients to write MCAP to targets
other than the filesystem.
### Docs

Detailed in docstrings. Will add an example before merge.

### Description

Today Rust clients can create MCAP writers that wrap an arbitrary
std::io::Write implementation of their own. This does not work in C++ or
in python. This PR adds that ability to the C++ SDK. This is useful for
anyone who want to log MCAP but not to a traditional filesystem - they
may want to log to a socket, or they may want to intercept writes before
they go to the filesystem.
<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

## TODO
- [x] Unit tests
- [x] Example code

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->